+++ navigation-experimental/src/main/kotlin/com/freeletics/khonshu/navigation/internal/MultiStack.kt
@@ -1,4 +1,4 @@
-package com.freeletics.khonshu.navigation.compose.internal
+package com.freeletics.khonshu.navigation.internal
 
 import android.os.Bundle
 import androidx.compose.runtime.MutableState
@@ -9,8 +9,6 @@ import com.freeletics.khonshu.navigation.BaseRoute
 import com.freeletics.khonshu.navigation.ContentDestination
 import com.freeletics.khonshu.navigation.NavRoot
 import com.freeletics.khonshu.navigation.NavRoute
-import com.freeletics.khonshu.navigation.internal.DestinationId
-import com.freeletics.khonshu.navigation.internal.destinationId
 import java.util.UUID
 import kotlinx.collections.immutable.ImmutableList
