+++ navigation-experimental/src/main/kotlin/com/freeletics/khonshu/navigation/internal/StackEntry.kt
@@ -1,11 +1,10 @@
-package com.freeletics.khonshu.navigation.compose.internal
+package com.freeletics.khonshu.navigation.internal
 
 import androidx.compose.runtime.Immutable
 import com.freeletics.khonshu.navigation.BaseRoute
 import com.freeletics.khonshu.navigation.ContentDestination
 import com.freeletics.khonshu.navigation.NavRoot
 import com.freeletics.khonshu.navigation.NavRoute
-import com.freeletics.khonshu.navigation.internal.destinationId
 import dev.drewhamilton.poko.Poko
 
 @Poko
